Pseudo-removed leave-out cross-validation

Pseudo-removed leave-out cross-validation

cv.pr.w(
  modern_taxa,
  modern_climate,
  nPLS = 5,
  trainfun,
  predictfun,
  pseudo,
  usefx = FALSE,
  fx = NA,
  cpus = 4,
  test_mode = TRUE,
  test_it = 5
)

Arguments

modern_taxa
the modern taxa abundance data, each row represents a sampling site, each column represents a taxon.

modern_climate
the modern climate value at each sampling site

nPLS
the number of components to be extracted

trainfun
training function you want to use, either WAPLS.w or TWAPLS.w

predictfun
predict function you want to use: if trainfun is WAPLS.w, then this should be WAPLS.predict.w; if trainfun is TWAPLS.w, then this should be TWAPLS.predict.w

pseudo
the geographically and climatically close sites to each test site, obtained from get_pseudo function

usefx
boolean flag on whether or not use fx correction.

fx
the frequency of the climate value for fx correction: if usefx is FALSE, this should be NA; otherwise, this should be obtained from the fx function.

cpus
number of CPUs for simultaneous iterations to execute, check parallel::detectCores() for available CPUs on your machine.

test_mode
boolean flag to execute the function with a limited number of iterations, test_it, for testing purposes only.

test_it
number of iterations to use in the test mode

Value

leave-one-out cross validation results

See also

Examples

if (FALSE) {
    # Load modern pollen data
    modern_pollen <- read.csv(system.file("extdata", 
                                          "Modern_Pollen_gdd_alpha_Tmin.csv", 
                                          package = "fxTWAPLS", 
                                          mustWork = TRUE))
    point <- modern_pollen[, c("Long", "Lat")]
    dist <- fxTWAPLS::get_distance(point, cpus = 1)
    pseudo_Tmin <- fxTWAPLS::get_pseudo(dist, 
                                        modern_pollen$Tmin, 
                                        cpus = 1)
    # Test WAPLS
    cv_pr_Tmin <- fxTWAPLS::cv.pr.w(taxa,
                                    modern_pollen$Tmin,
                                    nPLS = 5,
                                    fxTWAPLS::TWAPLS.w,
                                    fxTWAPLS::TWAPLS.predict.w,
                                    pseudo_Tmin,
                                    cpus = 1,
                                    test_mode = TRUE,
                                    test_it = test_it)
    # Test TWAPLS
    cv_pr_t_Tmin <- fxTWAPLS::cv.pr.w(taxa,
                                      modern_pollen$Tmin,
                                      nPLS = 5,
                                      fxTWAPLS::TWAPLS.w,
                                      fxTWAPLS::TWAPLS.predict.w,
                                      pseudo_Tmin,
                                      cpus = 1, 
                                      test_mode = TRUE,
                                      test_it = test_it)
}