Pseudo removed leave out cross validation

Pseudo removed leave out cross validation

cv.pr.w(
  modern_taxa,
  modern_climate,
  nPLS = 5,
  trainfun,
  predictfun,
  pseudo,
  usefx = FALSE,
  fx = NA,
  cpus = 4
)

Arguments

modern_taxa
the modern taxa abundance data, each row represents a sampling site, each column represents a taxon.

modern_climate
the modern climate value at each sampling site

nPLS
the number of components to be extracted

trainfun
training function you want to use, either WAPLS.w or TWAPLS.w

predictfun
predict function you want to use: if trainfun is WAPLS.w, then this should be WAPLS.predict.w; if trainfun is TWAPLS.w, then this should be TWAPLS.predict.w

pseudo
the geographically and climatically close sites to each test site, obtained from get_pseudo function

usefx
boolean flag on whether or not use fx correction.

fx
the frequency of the climate value for fx correction: if usefx is FALSE, this should be NA; otherwise, this should be obtained from the fx function.

cpus
number of CPUs for simultaneous iterations to execute, check parallel::detectCores() for available CPUs on your machine.

Value

leave-one-out cross validation results