Random t-test

Do a random t-test to the cross-validation results.

rand.t.test.w(cvoutput, n.perm = 999)

Arguments

cvoutput
Cross-validation output either from cv.w or cv.pr.w.

n.perm
The number of permutation times to get the p value, which assesses whether using the current number of components is significantly different from using one less.

Value

A matrix of the statistics of the cross-validation results. Each component is described below:

R2

the coefficient of determination (the larger, the better the fit).

Avg.Bias

average bias.

Max.Bias

maximum bias.

Min.Bias

minimum bias.

RMSEP

root-mean-square error of prediction (the smaller, the better the fit).

delta.RMSEP

the percent change of RMSEP using the current number of components than using one component less.

p

assesses whether using the current number of components is significantly different from using one component less, which is used to choose the last significant number of components to avoid over-fitting.

-

The degree of overall compression is assessed by doing linear regression to the cross-validation result and the observed climate values.

Compre.b0: the intercept.

Compre.b1: the slope (the closer to 1, the less the overall compression).

Compre.b0.se: the standard error of the intercept.

Compre.b1.se: the standard error of the slope.

See also

Examples

if (FALSE) {
# Load modern pollen data
modern_pollen <- read.csv("/path/to/modern_pollen.csv")
                                      
# Extract taxa
taxaColMin <- which(colnames(modern_pollen) == "taxa0")
taxaColMax <- which(colnames(modern_pollen) == "taxaN")
taxa <- modern_pollen[, taxaColMin:taxaColMax]

# Get the frequency of each climate variable fx
fx_Tmin <- fxTWAPLS::fx(modern_pollen$Tmin, bin = 0.02)
fx_gdd <- fxTWAPLS::fx(modern_pollen$gdd, bin = 20)
fx_alpha <- fxTWAPLS::fx(modern_pollen$alpha, bin = 0.002)

# MTCO
## fx
fit_Tmin <- fxTWAPLS::WAPLS.w(taxa, modern_pollen$Tmin, nPLS = 5)
## LOOCV
test_mode <- TRUE # It should be set to FALSE before running
### without fx
cv_Tmin <- fxTWAPLS::cv.w(taxa,
                          modern_pollen$Tmin,
                          nPLS = 5,
                          fxTWAPLS::WAPLS.w,
                          fxTWAPLS::WAPLS.predict.w,
                          cpus = 2, # Remove the following line
                          test_mode = test_mode)
### with fx
cv_f_Tmin <- fxTWAPLS::cv.w(taxa,
                            modern_pollen$Tmin,
                            nPLS = 5,
                            fxTWAPLS::WAPLS.w,
                            fxTWAPLS::WAPLS.predict.w,
                            usefx = TRUE,
                            fx = fx_Tmin,
                            cpus = 2, # Remove the following line
                            test_mode = test_mode)
                            
## Random t-test
rand_Tmin <- fxTWAPLS::rand.t.test.w(cv_Tmin, n.perm = 999)
rand_f_Tmin <- fxTWAPLS::rand.t.test.w(cv_f_Tmin, n.perm = 999)
}